Improved cursor loader.
The new cursor loader is now able to fix cursor images so the host system is able to work with them properly. It performs changes of the texture size and is able to fix alpha values. It works with cursors that are too small for the host system and with cursor too large. How ever the support for larger cursors is limited to avoid crashes. The cursor will likely look crappy.